Workaround: comment out ndk-glue dev-dependency
For some reason Cargo is making ndk-glue an inherited dependency for
crates that are dependent on cpal. In this case though ndk-glue doesn't
just end up as a benign/redundant dependency, it can also cause the
application to crash on load.

As far as I can tell this seems like a Cargo bug considering that the
Cargo reference manual clearly documents:

> Dev-dependencies are not used when compiling a package for building,
> but are used for compiling tests, examples, and benchmarks.
>
> These dependencies are not propagated to other packages which depend
> on this package.

.. and we're observing a crash building an Android application dependent
on cpal, due to it's dev-dependency on ndk-glue propagating.

As a workaround it works to just comment this dev-dependency out since
it's only there for the optional android example.